requirements.txt
python-dotenv
langchain
openai
tiktoken
PyPDF2
faiss-cpu
langchain-openai
streamlit
streamlit-chat
